# specs/openapi/adscenter.yaml
openapi: 3.0.3
info:
  title: Adscenter Service
  version: 0.1.0
  license:
    name: Elastic-Internal
servers:
  - url: /
paths:
...
  /api/v1/adscenter/diagnose:
    post:
      operationId: diagnose
      summary: Diagnose ad account issues based on provided metrics and landing URL
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                accountId: { type: string }
                landingUrl: { type: string }
                metrics:
                  type: object
                  additionalProperties: true
              required: [accountId]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiagnoseResult'
        '400': { description: Bad Request }
        '401': { description: Unauthorized }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DiagnoseRule:
      type: object
      properties:
        code: { type: string }
        severity: { type: string, enum: [info, warn, error] }
        message: { type: string }
        details:
          type: object
          additionalProperties: true
      required: [code, severity, message]
    SuggestedAction:
      type: object
      properties:
        action: { type: string }
        params:
          type: object
          additionalProperties: true
        reason: { type: string }
      required: [action]
    DiagnoseResult:
      type: object
      properties:
        summary: { type: string, enum: [ok, warn, error] }
        rules:
          type: array
          items: { $ref: '#/components/schemas/DiagnoseRule' }
        suggestedActions:
          type: array
          items: { $ref: '#/components/schemas/SuggestedAction' }
      required: [summary, rules]
...